Renamed factory_proposal.py to factory.py
