+++ ctapipe/calib/camera/charge_extractors.py
@@ -3,7 +3,7 @@ import numpy as np
 from traitlets import Int, CaselessStrEnum
 
 from ctapipe.core import Component
-from ctapipe.calib.camera.factory_proposal import Factory
+from ctapipe.calib.camera.factory import Factory
 
 
 class ChargeExtractor(Component):
