+++ ctapipe/calib/camera/factory.py
